Move samplers to util package
